🎨 Refactor application structure by moving App class to a new directory and updating includes
